add Marketplace feature to HomeTabs navigation
This update introduces the Marketplace view for both farmers and consumers. The MarketScreen was imported and added to a new MarketStack with a Stack.Screen for navigation. The Marketplace was integrated into the HomeTabs using a Tab.Screen with the title "Marketplace" and assigned an appropriate index and icon (storefront-outline) for clear representation. Note that features like viewing product details and direct contact options are not yet included in this release.